board split by normal and item
