+++ src/seoultech/se/tetris/main/Tetris.java
@@ -1,11 +1,12 @@
 package seoultech.se.tetris.main;
 
-import seoultech.se.tetris.component.board.ItemBoard;
+import seoultech.se.tetris.component.board.*;
 
 public class Tetris {
 
 	public static void main(String[] args) {
-		ItemBoard main = new ItemBoard();
+		Board main = new ItemBoard();
+//		Board main = new NormalBoard();
 
 		main.setSize(400, 600);
 		main.setLocation(1700, 0);
